requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
flask>=2.0.0
flask-session>=0.5.0
python-dotenv>=1.0.0
pymongo>=4.5.0
requests>=2.31.0
pyyaml>=6.0
